# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Doação de Livros",
    "description": "API para gerenciamento de doações de livros, incluindo cadastro de usuários, cadastro de livros, e controle de doações.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "paths": {
    "/cadastrar_usuario": {
      "post": {
        "summary": "Cadastrar um novo usuário",
        "description": "Endpoint para cadastro de novos usuários no sistema. O sistema gera automaticamente um ID para o usuário.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nome", "email"],
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "João Silva",
                    "description": "Nome completo do usuário"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "joao.silva@email.com",
                    "description": "Email do usuário (deve ser único)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário cadastrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Usuário cadastrado com sucesso"
                    },
                    "id": {
                      "type": "integer",
                      "example": 1,
                      "description": "ID do usuário gerado automaticamente (auto-incremento)"
                    },
                    "nome": {
                      "type": "string",
                      "example": "João Silva"
                    },
                    "email": {
                      "type": "string",
                      "example": "joao.silva@email.com"
                    },
                    "data_cadastro": {
                      "type": "string",
                      "example": "28/06/2026",
                      "description": "Data de cadastro no formato DD/MM/AAAA"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro na requisição - campos obrigatórios faltando ou email já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": {
                      "type": "string",
                      "example": "Email já cadastrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/buscar_usuario/{usuario_id}": {
      "get": {
        "summary": "Buscar usuário por ID",
        "description": "Endpoint para buscar informações de um usuário específico através do seu ID.",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "description": "ID do usuário",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "nome": {
                      "type": "string",
                      "example": "João Silva"
                    },
                    "email": {
                      "type": "string",
                      "example": "joao.silva@email.com"
                    },
                    "data_cadastro": {
                      "type": "string",
                      "example": "28/06/2026"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": {
                      "type": "string",
                      "example": "Usuário não encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/cadastrar_livro": {
      "post": {
        "summary": "Cadastrar um novo livro para doação",
        "description": "Endpoint para cadastro de novos livros disponíveis para doação. O livro é associado ao usuário doador através do ID do usuário.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["titulo", "autor", "usuario_doador_id"],
                "properties": {
                  "titulo": {
                    "type": "string",
                    "example": "O Pequeno Príncipe",
                    "description": "Título do livro"
                  },
                  "autor": {
                    "type": "string",
                    "example": "Antoine de Saint-Exupéry",
                    "description": "Nome do autor do livro"
                  },
                  "usuario_doador_id": {
                    "type": "integer",
                    "example": 1,
                    "description": "ID do usuário doador"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Livro cadastrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Livro cadastrado com sucesso"
                    },
                    "id": {
                      "type": "integer",
                      "example": 1,
                      "description": "ID do livro gerado automaticamente"
                    },
                    "titulo": {
                      "type": "string",
                      "example": "O Pequeno Príncipe"
                    },
                    "autor": {
                      "type": "string",
                      "example": "Antoine de Saint-Exupéry"
                    },
                    "usuario_doador_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "data_cadastro": {
                      "type": "string",
                      "example": "28/06/2026"
                    },
                    "status": {
                      "type": "string",
                      "example": "disponivel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro na requisição - campos obrigatórios faltando ou usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": {
                      "type": "string",
                      "example": "Usuário não encontrado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": {
                      "type": "string",
                      "example": "Usuário não encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/livros_disponiveis": {
      "get": {
        "summary": "Listar todos os livros disponíveis para doação",
        "description": "Endpoint para listar todos os livros que estão disponíveis para doação (status: disponivel).",
        "responses": {
          "200": {
            "description": "Lista de livros disponíveis retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "example": 5,
                      "description": "Total de livros disponíveis"
                    },
                    "livros": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "titulo": {
                            "type": "string",
                            "example": "O Pequeno Príncipe"
                          },
                          "autor": {
                            "type": "string",
                            "example": "Antoine de Saint-Exupéry"
                          },
                          "usuario_doador_id": {
                            "type": "integer",
                            "example": 1
                          },
                          "data_cadastro": {
                            "type": "string",
                            "example": "28/06/2026"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/doar_livro": {
      "post": {
        "summary": "Realizar a doação de um livro",
        "description": "Endpoint para realizar a doação de um livro disponível para um usuário receptor. Atualiza o status do livro para 'doado' e registra a data da doação.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["livro_id", "usuario_receptor_id"],
                "properties": {
                  "livro_id": {
                    "type": "integer",
                    "example": 1,
                    "description": "ID do livro a ser doado"
                  },
                  "usuario_receptor_id": {
                    "type": "integer",
                    "example": 2,
                    "description": "ID do usuário que receberá o livro"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Doação realizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Doação realizada com sucesso"
                    },
                    "livro_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "titulo": {
                      "type": "string",
                      "example": "O Pequeno Príncipe"
                    },
                    "usuario_doador_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "usuario_receptor_id": {
                      "type": "integer",
                      "example": 2
                    },
                    "data_doacao": {
                      "type": "string",
                      "example": "28/06/2026"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro na requisição - campos obrigatórios faltando ou livro não disponível",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": {
                      "type": "string",
                      "example": "Livro não está disponível para doação"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Livro ou usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": {
                      "type": "string",
                      "example": "Usuário receptor não encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/deletar_livro/{livro_id}": {
      "delete": {
        "summary": "Deletar um livro",
        "description": "Endpoint para deletar um livro do sistema. Só é possível deletar livros que estão disponíveis (não doados).",
        "parameters": [
          {
            "name": "livro_id",
            "in": "path",
            "required": true,
            "description": "ID do livro a ser deletado",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Livro deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Livro deletado com sucesso"
                    },
                    "id": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro na requisição - livro já foi doado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": {
                      "type": "string",
                      "example": "Não é possível deletar livro que já foi doado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Livro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erro": {
                      "type": "string",
                      "example": "Livro não encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/usuarios": {
      "get": {
        "summary": "Listar todos os usuários cadastrados",
        "description": "Endpoint para listar todos os usuários cadastrados no sistema.",
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "example": 10,
                      "description": "Total de usuários cadastrados"
                    },
                    "usuarios": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "nome": {
                            "type": "string",
                            "example": "João Silva"
                          },
                          "email": {
                            "type": "string",
                            "example": "joao.silva@email.com"
                          },
                          "data_cadastro": {
                            "type": "string",
                            "example": "28/06/2026"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {}
  }
}